Moves find_functions to graph_utils.py
